Use adapter structure for reviews
