+++ src/adapters/v01/listings_adapter.js
@@ -25,6 +25,11 @@ const fractionalListingSchemaId = 'fractional-listing.json'
 const unitPurchaseSchemaId = 'unit-purchase.json'
 const fractionalPurchaseSchemaId = 'fractional-purchase.json'
 
+const buyerReviewStage = 4
+const sellerReviewStage = 5
+const reviewStages = [buyerReviewStage, sellerReviewStage]
+const emptyIpfsHash = '0x0000000000000000000000000000000000000000000000000000000000000000'
+
 const ajv = new Ajv({
   schemas: [
     listingSchema,
@@ -295,6 +300,53 @@ class ListingsAdapter {
     )
     await Promise.all(fetchPromises)
     return mapped
+  }
+
+  async findReviewForPurchase(contractService, ipfsService, listingId, purchaseId) {
+    const v01_ListingsContract = await contractService.deployed(
+      contractService.v01_ListingsContract
+    )
+    const logs = await new Promise((resolve, reject) => {
+      v01_ListingsContract.getPastEvents(
+        'PurchaseChange',
+        {
+          fromBlock: 0,
+          toBlock: 'latest',
+          filter: { _listingIndex: listingId, _purchaseIndex: purchaseId }
+        },
+        function(error, logs) {
+          if (error) {
+            reject(error)
+          }
+        resolve(logs)
+      })
+    })
+    const createItem = async (log) => {
+      const asBytes32 = log.returnValues._ipfsHash
+      const stage = log.returnValues._stage
+      const ipfsHash = contractService.getIpfsHashFromBytes32(asBytes32)
+      let ipfsData
+      if (ipfsHash !== emptyIpfsHash) {
+        ipfsData = await ipfsService.getFile(ipfsHash)
+      }
+      return { ipfsData, stage }
+    }
+    const itemsPromise = logs.filter(log => {
+      return reviewStages.includes(Number(log.returnValues._stage))
+    }).map(createItem)
+    const items = await Promise.all(itemsPromise)
+    const buyerReviewItems = items.filter(({ stage }) => {
+      return Number(stage) === buyerReviewStage
+    })
+    const sellerReviewItems = items.filter(({ stage }) => {
+      return Number(stage) === sellerReviewStage
+    })
+    const format = (item) => {
+      return { ipfsData: item.ipfsData }
+    }
+    const fromBuyer = buyerReviewItems.length ? format(buyerReviewItems[0]) : null
+    const fromSeller = sellerReviewItems.length ? format(sellerReviewItems[0]) : null
+    return { fromBuyer, fromSeller }
   }
 }
